feat: started work on removing Molecule
